Remove unmaintained Showcase section
It's too limited in features and doesn't let us properly put forward
the most showcase-worthy projects, so it ends up giving a bad impression
of the engine.

We can reintroduce a showcase if we do things right, with custom development
to make a good-looking, easy to maintain and attractive showcase.

Fixes #7.