Add a dark theme for the website, overhaul usability and accessibility
Like for the documentation, this theme is used depending on the
OS/browser's preference. On Firefox, the Dark Website Forcer add-on
can be installed to use a dark theme if the OS isn't configured to
use a dark theme.

This commit also improves hover/active feedback by using the CSS
`filter` property. This behaves better with images and buttons
without having to manually choose colors for every interaction.

This commit includes miscellaneous changes to the homepage design
(such as the "A different way to make games" heading now being
left-aligned instead of right-aligned).

A visible outline is now added to focused elements for better
keyboard navigation.

Navigation bar links now have padding to be easier to click,
both on desktop and on mobile.