Add October CacheRoute plugin to speed up page loading
This provides a significant speedup by caching the CMS' pages.
This way, the database no longer needs to be queried and templates
no longer need to be rendered to display pages.

There's no need to bother setting up the plugin for development.

Recommended configuration for production:

| Route pattern | TTL (Time To Live)      |
|---------------|-------------------------|
| backend/*     | 0 minutes               |
| download      | 0 minutes               |
| /             | 5 minutes               |
| article/*     | 5 minutes               |
| news*         | 5 minutes               |
| devblog*      | 5 minutes               |
| *             | 525600 minutes (1 year) |

Static pages are cached for a long duration, pages featuring dynamic
(blog) content are cached for a short duration, and backend pages
are not cached at all. The download per-OS redirect is also not
cached either as that causes issues; only the OS-specific download
pages are cached.

Make sure to run `php artisan plugin:refresh serenitynow.cacheroute`
in the October folder after deploying the plugin.

Remember to clear the cache after a deployment by running
`php artisan cache:clear` in the October folder.