More grammatically proper headline
"The game engine you waited for" is a bit awkward in English, it's all in past tense as though the addressee once was looking for an engine like this but is not necessarily seeking one right now. The proposed change is a more typical way of saying what I believe the original author meant. You can google the idiom and will find many examples acknowledging it, here is one of the first that pops up:  https://www.gymglish.com/en/gymglish/english-translation/the-moment-youve-been-waiting-for