Teams: Add Production team
It's been operating for a while but was not part of the main draft for
the Teams page, and we weren't sure initially if it should be included
since it uses a private channel for communication.

But it would be good for contributors to know who to reach out to for
production questions, and they can use the devel channel for this.