Fix incorrect link on preview downloads page
There's a link at the top of the [previews download page](https://godotengine.org/download/preview/) which says "Looking for a stable version? See below!"

This link is supposed to scroll the page to the section lower down where you can download the stable version. However, prior to this commit it didn't work because the link pointed to an anchor that didn't exist.

The link pointed to `#stable`. However, the actual anchor, as defined in `download-section.html`, is `#call-to-action`.

This commit fixes the link so that clicking it works properly.