Add custom anchor to download-section.html, and use it to fix link on previews download page (#1000)
* Add custom anchor to download-section.html, and use it to fix link on preview.html

There's a link at the top of the previews download page which says "Looking for a stable version? See below!"

This link is supposed to scroll the page to the section lower down where you can download the stable version. However, prior to this commit, the link doesn't work because it points to an anchor that no longer exists.

The link points to #stable. However, the actual anchor, as defined in download-section.html, is #call-to-action.

This commit lets you set a custom anchor when you include download-section.html, and does so in preview.html to fix the broken link.

* Simplify header in download-section.html